+++ src/EriadorMapWindow.lua
@@ -97,6 +97,9 @@ function EriadorMapWindow:Constructor(parent, class, race, shortcuts)
         {-1, -1}, -- Andrath
         {-1, -1}, -- Swanfleet
         {-1, -1}, -- Cardolan
+        {-1, -1}, -- Carn Dûm
+        {-1, -1}, -- Clegur
+        {-1, -1}, -- Pelargir
 
         -- Warden skills
         {725, 400}, -- Ost Guruth
@@ -144,6 +147,25 @@ function EriadorMapWindow:Constructor(parent, class, race, shortcuts)
         {-1, -1}, -- Andrath
         {-1, -1}, -- Swanfleet
         {-1, -1}, -- Cardolan
+        {-1, -1}, -- Carn Dûm
+        {-1, -1}, -- Clegur
+        {-1, -1}, -- Pelargir
+
+        -- Mariner skills
+        {-1, -1}, -- Celondim
+        {-1, -1}, -- Brown Lands
+        {-1, -1}, -- Trader's Wharf
+        {-1, -1}, -- after-battle Osgiliath
+        {-1, -1}, -- Lake-town
+        {-1, -1}, -- Mirk-eaves
+        {-1, -1}, -- Tinnundir
+        {-1, -1}, -- Dol Amroth
+        {-1, -1}, -- Buckland
+        {-1, -1}, -- Pelegir
+        {-1, -1}, -- Sûri-kylä
+        {-1, -1}, -- Lothlórien
+        {-1, -1}, -- Tharbad
+        {-1, -1}, -- Snowbourn
 
         -- Reputation skills
         {305, 330}, -- Thorin's Hall
@@ -196,6 +218,10 @@ function EriadorMapWindow:Constructor(parent, class, race, shortcuts)
         {-1, -1}, -- Swanfleet
         {-1, -1}, -- Cardolan
         {-1, -1}, -- Tracery Archive
+        {-1, -1}, -- Carn Dûm
+        {-1, -1}, -- Clegur
+        {-1, -1}, -- Glân Vraig
+        {-1, -1}, -- Pelargir
 
         -- Racial skills
         {620, 400}, -- Bree
@@ -205,6 +231,7 @@ function EriadorMapWindow:Constructor(parent, class, race, shortcuts)
         {-1, -1},   -- Grimbeorn's House
         {-1, -1},   -- Caras Galadhon
         {330, 330}, -- Thorin's Gate
+        {-1, -1},   -- Lyndelby
         -- Moors Map
         {815, 300} -- Ettenmoors
     };
@@ -248,6 +275,18 @@ function EriadorMapWindow:AddShortcuts()
         if (self.mapLocations[counter][1] > 0 and self.PlayerClass == Turbine.Gameplay.Class.Warden and
             self:IsShortcutEnabled(wardenLocations:IdAtIndex(i)) == 1) then
             self:AddSingleShortcut(counter, Turbine.UI.Lotro.Shortcut(6.0, wardenLocations:IdAtIndex(i)));
+        end
+
+        counter = counter + 1;
+    end
+
+    -- mariner locations
+    for i = 1, travelCount[7], 1 do
+
+        -- add mariner locations if they should be on this map
+        if (self.mapLocations[counter][1] > 0 and self.PlayerClass == Turbine.Gameplay.Class.Corsair and
+            self:IsShortcutEnabled(marinerLocations:IdAtIndex(i)) == 1) then
+            self:AddSingleShortcut(counter, Turbine.UI.Lotro.Shortcut(6.0, marinerLocations:IdAtIndex(i)));
         end
 
         counter = counter + 1;