+++ src/IndexedDictionaryEn.lua
@@ -74,9 +74,10 @@ function TravelDictionaries:Constructor()
     racialLocations = IndexedDictionary();
     repLocations = IndexedDictionary();
     creepLocations = IndexedDictionary();
+    marinerLocations = IndexedDictionary();
 
     -- a table to store the number of each type of travel skill
-    travelCount = { 0, 0, 0, 0, 0, 0 };
+    travelCount = { 0, 0, 0, 0, 0, 0, 0 };
     LocationsTotal = 0;
 
     self:CreateDictionaries();
@@ -91,6 +92,9 @@ function TravelDictionaries:GetCounts()
     travelCount[4] = repLocations:GetCount();
     travelCount[5] = racialLocations:GetCount();
     travelCount[6] = creepLocations:GetCount();
+    travelCount[7] = marinerLocations:GetCount();
+
+    LocationsTotal = travelCount[1] + travelCount[2] + travelCount[3] + travelCount[4] + travelCount[5] + travelCount[6] + travelCount[7];
 end
 
 -- function to return the count of a specific type of travel
@@ -98,13 +102,19 @@ function TravelDictionaries:GetCount(index)
     return travelCount[index];
 end
 
+-- function to get the total number of travel skills
+function TravelDictionaries:GetTotalCount()
+    return LocationsTotal;
+end
+
 -- function to return the 5 tables containing the travel skills
 function TravelDictionaries:GetLocations()
-    return genLocations, hunterLocations, wardenLocations, racialLocations, repLocations;
+    return genLocations, hunterLocations, wardenLocations, racialLocations, repLocations, creepLocations, marinerLocations;
 end
 
 -- add the data to custom dictionaries to maintain the order
 function TravelDictionaries:CreateDictionaries()
+    -- add the hunter locations
     hunterLocations:AddSkill("Return to Camp", "0x7000A2C1", "Return to Camp");
     hunterLocations:AddSkill("Guide to Bree", "0x70003F42", "Bree (Guide)");
     hunterLocations:AddSkill("Guide to Thorin's Hall", "0x70003F41", "Thorin's Hall (Guide)");
@@ -161,6 +171,7 @@ function TravelDictionaries:CreateDictionaries()
     hunterLocations:AddSkill("Guide to Clegur", "0x70064F4C", "Clegur (Guide)");
     hunterLocations:AddSkill("Guide to Pelargir", "0x700658EA", "Pelargir (Guide)");
 
+    -- add the Warden locations
     wardenLocations:AddSkill("Muster in Ost Guruth", "0x70014786", "Ost Guruth (Muster)");
     wardenLocations:AddSkill("Muster in Esteldín", "0x70014798", "Esteldín (Muster)");
     wardenLocations:AddSkill("Muster in Evendim", "0x7001478E", "Evendim (Muster)");
@@ -210,14 +221,31 @@ function TravelDictionaries:CreateDictionaries()
     wardenLocations:AddSkill("Muster at Clegur", "0x70064F4D", "Clegur (Muster)");
     wardenLocations:AddSkill("Muster at Pelargir", "0x700658E8", "Pelargir (Muster)");
 
+    -- add the Mariner locations
+    marinerLocations:AddSkill("Sail to Celondim", "0x70066100", "Celondim (Sail)")
+    marinerLocations:AddSkill("Sail to the Brown Lands", "0x70066101", "Brown Lands (Sail)")
+    marinerLocations:AddSkill("Sail to Trader's Wharf", "0x70066105", "Trader's Wharf (Sail)")
+    marinerLocations:AddSkill("Sail to after-battle Osgiliath", "0x70066109", "After-battle Osgiliath (Sail)")
+    marinerLocations:AddSkill("Sail to Lake-town", "0x7006610C", "Lake-town (Sail)")
+    marinerLocations:AddSkill("Sail to the Mirk-eaves", "0x7006610E", "Mirk-eaves (Sail)")
+    marinerLocations:AddSkill("Sail to Tinnudir", "0x7006610F", "Tinnudir (Sail)")
+    marinerLocations:AddSkill("Sail to Dol Amroth", "0x70066117", "Dol Amroth (Sail)")
+    marinerLocations:AddSkill("Sail to Buckland", "0x7006611A", "Buckland (Sail)")
+    marinerLocations:AddSkill("Sail to Pelargir", "0x7006611B", "Pelargir (Sail)")
+    marinerLocations:AddSkill("Sail to Sûri-kylä", "0x7006611C", "Sûri-kylä (Sail)")
+    marinerLocations:AddSkill("Sail to Lothlórien", "0x7006611E", "Lothlórien (Sail)")
+    marinerLocations:AddSkill("Sail to Tharbad", "0x70066120", "Tharbad (Sail)")
+    marinerLocations:AddSkill("Sail to Snowbourn", "0x70066121", "Snowbourn (Sail)")
+
     -- add the racial travel skills
-    racialLocations:AddSkill("Return to Bree", "0x700062F6", "Bree (Racial)");
-    racialLocations:AddSkill("Return to the Shire", "0x700062C8", "Shire (Racial)");
+    racialLocations:AddSkill("Return to Bree", "0x700062F6", "Bree (Racial)"); -- human
+    racialLocations:AddSkill("Return to the Shire", "0x700062C8", "Shire (Racial)"); -- hobbit
     racialLocations:AddSkill("Return to Thorin's Gate", "0x70006346", "Thorin's Gate (Racial)"); -- dwarf
-    racialLocations:AddSkill("Return to Rivendell", "0x7000631F", "Rivendell (Racial)");
-    racialLocations:AddSkill("Return Home", "0x70041A22", "Beorning Home (Racial)");
-    racialLocations:AddSkill("Travel to Caras Galadhon in Lothlórien", "0x70048C8C", "Caras Galadhon (Racial)");
+    racialLocations:AddSkill("Return to Rivendell", "0x7000631F", "Rivendell (Racial)"); --elf
+    racialLocations:AddSkill("Return Home", "0x70041A22", "Beorning Home (Racial)"); -- beorning
+    racialLocations:AddSkill("Travel to Caras Galadhon in Lothlórien", "0x70048C8C", "Caras Galadhon (Racial)"); -- high-elf
     racialLocations:AddSkill("Travel to Thorin's Hall", "0x70053C0F", "Thorin's Hall (Racial)"); -- stoutaxe
+    racialLocations:AddSkill("Return to Lyndelby", "0x70066D31", "Lyndelby (Racial)") -- river hobbit
 
     -- add the Return Home and housing skills
     genLocations:AddSkill("Return Home",    "0x700256BA", "Return Home");